requirements.txt


pydantic>=2.5.1
typing-extensions>=4.7.1
python-dotenv>=1.1.1


langgraph>=0.1.0


openai>=1.1.0


pdfplumber>=0.9.0


pytest>=7.4.0